# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pet Spa Booking API Documentation",
    "description": "RESTful API cho hệ thống đặt lịch Pet Spa",
    "version": "1.0.0",
    "contact": {
      "name": "Pet Spa Support",
      "email": "admin@petspa.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Pet Spa Booking API Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Bearer token"
      }
    },
...
  },
  "paths": {
...
    "/admin/dashboard": {
      "get": {
        "tags": ["Admin Dashboard"],
        "summary": "Thống kê tổng quan dashboard",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Thành công"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden - Admin only"}
        }
      }
    },
    "/admin/dashboard/revenue": {
      "get": {
        "tags": ["Admin Dashboard"],
        "summary": "Thống kê doanh thu",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "start_date", "in": "query", "schema": {"type": "string", "format": "date"}},
          {"name": "end_date", "in": "query", "schema": {"type": "string", "format": "date"}}
        ],
        "responses": {
          "200": {"description": "Thành công"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden - Admin only"}
        }
      }
    },
...
    "/admin/appointments": {
      "get": {
        "tags": ["Admin Appointments"],
        "summary": "Lấy danh sách tất cả lịch hẹn",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Thành công"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden - Admin only"}
        }
      }
    },
    "/admin/appointments/{id}": {
      "get": {
        "tags": ["Admin Appointments"],
        "summary": "Lấy chi tiết lịch hẹn",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {
          "200": {"description": "Thành công"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Không tìm thấy"}
        }
      },
      "delete": {
        "tags": ["Admin Appointments"],
        "summary": "Xóa lịch hẹn",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {
          "200": {"description": "Xóa thành công"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Không tìm thấy"}
        }
      }
    },
    "/admin/appointments/{id}/status": {
      "patch": {
        "tags": ["Admin Appointments"],
        "summary": "Cập nhật trạng thái lịch hẹn",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": {"type": "string", "enum": ["Pending", "Confirmed", "Completed", "Cancelled"]}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Cập nhật thành công"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Không tìm thấy"}
        }
      }
    },
...
  }
}